final/requirements.txt
Flask==0.12.2
Werkzeug<0.13.0,>=0.12.0
requests
passlib
typing
grpcio
cachetools
pyasn1
pyasn1_modules
rsa
gunicorn
google-cloud-language

